[renderer-ff] Change the branching point to the Renderer
