+++ packages/pure-markdown/src/index.ts
@@ -1,19 +1,19 @@
 import SimpleMarkdown from "@khanacademy/simple-markdown";
 
 import {pureMarkdownRules} from "./rules";
-import {pureMarkdownRules as rulesWithRendererUpgrade} from "./rules.new";
+import {pureMarkdownRulesNew} from "./rules.new";
 
 export {traverseContent} from "./traversal";
 export {libVersion} from "./version";
 
 export {pureMarkdownRules};
-export {rulesWithRendererUpgrade};
+export {pureMarkdownRulesNew};
 
 // @ts-expect-error - TS2345 - Argument of type '{ readonly columns: { readonly order: -2; readonly match: any; readonly parse: (capture: any, parse: any, state: any) => any; }; readonly crowdinId: { readonly order: -1; readonly match: (source: any, state: any, prevCapture: any) => any; readonly parse: (capture: any, parse: any, state: any) => any; }; ... 34 more ...' is not assignable to parameter of type 'ParserRules'.
 const builtParser = SimpleMarkdown.parserFor(pureMarkdownRules);
 const builtParserWithRendererUpgrade = SimpleMarkdown.parserFor(
     // @ts-expect-error - TS2345 - Argument of type '{ readonly columns: { readonly order: -2; readonly match: any; readonly parse: (capture: any, parse: any, state: any) => any; }; readonly crowdinId: { readonly order: -1; readonly match: (source: any, state: any, prevCapture: any) => any; readonly parse: (capture: any, parse: any, state: any) => any; }; ... 34 more ...' is not assignable to parameter of type 'ParserRules'.
-    rulesWithRendererUpgrade,
+    pureMarkdownRulesNew,
 );
 /**
  * Parses a **Perseus** Markdown string into an AST.
